marker to forward msg per mail configurable
with the new flag '-forward' it's now possible to configure
a marker and receiver mapping.

example:
--------

  ./matterbot ... -forward user=user@mail.com ...

if the chat-message contains a '@user' prefix, the message
are forwarded to 'user@mail.com'

---------------------------------------------------------

to define more than one forward mapping, separate each
mapping with a comma.

it's expected that the marker in the chat-message are separated
witch space or comma.

example:
--------

start matterbot with:

  ./matterbot ... -forward user1=user1@mail.com,user2=abc@example.com ...

send a chat message:

  @user1, @inter @abc this is a chat message, which will be
  forwarded to 'user1@mail.com' and 'user2@example.com'